start integrations with keylogger and client
the client has the ability to start and end the keylogging. An int
in keylogger.c serves as an indicator for when to stop logging.
I haven't been able to check if the client can mutate the value
because the server is broken.

main now only has access to client.c, and through functions in client.c
it can start/stop the keylogging

still need to ensure the keylogging will stop when the socket stops though.